% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build-tex.r
\name{make_cvtex}
\alias{make_cvtex}
\title{Make vita.}
\usage{
make_cvtex(bib = "~/Dropbox/web/vitaKeane/keane.bib",
  path = "~/Dropbox/web/vitaWeb", templates = "~/Dropbox/web/vitaKeane",
  texName = "keane-cv.tex", clean = FALSE, compile = TRUE,
  texBin = "/Library/TeX/texbin/lualatex")
}
\description{
Make vita.
}

