% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.r
\name{make_vita}
\alias{make_vita}
\title{Make vita.}
\usage{
make_vita(bib = "~/Dropbox/web/vitaKeane/keane.bib",
  path = "~/Dropbox/web/vitaWeb", templates = "~/Dropbox/web/vita",
  clean = FALSE)
}
\description{
Make vita.
}

